[0.2.7] [WIP] Continuing path_tree_print() implementation:
- started implementation of the recursion loop of path_tree_print();
- implemented initial throwaway memory setup and allocating appropriate buffers on the first recursion pass;
- finally changed the incorrect type of capacity in memory_create (was long long, now size_t).

+++ include/memory.h
@@ -17,7 +17,7 @@ typedef struct
    void* pointer;
 } Memory;
 
-extern Memory memory_create(long long capacity);
+extern Memory memory_create(size_t capacity);
 extern void* memory_allocate(Memory* memory, size_t alloc_size);
 extern void memory_clear(Memory* memory);
 extern void memory_usage_status(Memory* memory);
